refactor: 경험 추가 자료 상태 로직 분리
